requirements.txt
discord.py==1.7.3
requests==2.27.1
